Fix APIs parmeters andi CallConnection Live tests.
