+++ sdk/communication/Azure.Communication.CallingServer/tests/CallConnection/CallConnectionTests.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Net;
 using System.Threading.Tasks;
 using NUnit.Framework;
@@ -12,14 +13,8 @@ namespace Azure.Communication.CallingServer.Tests
     public class CallConnectionTests : CallingServerTestBase
     {
         private const string CancelAllMediaOperaionsResponsePayload = "{" +
-                                                "\"operationId\": \"dummyId\"," +
                                                 "\"status\": \"completed\"," +
                                                 "\"operationContext\": \"dummyOperationContext\"," +
-                                                "\"resultInfo\": {" +
-                                                "\"code\": 200," +
-                                                "\"subcode\": 200," +
-                                                "\"message\": \"dummyMessage\"" +
-                                                  "}" +
                                                 "}";
 
         private const string PlayAudioResponsePayload = "{" +
@@ -35,6 +30,84 @@ namespace Azure.Communication.CallingServer.Tests
 
         private const string AddParticipantResultPayload = "{" +
                                                                 "\"participantId\": \"dummyparticipantid\"" +
+                                                            "}";
+
+        private const string TransferCallResultPayload = "{" +
+                                                "\"operationId\": \"dummyId\"," +
+                                                "\"status\": \"running\"," +
+                                                "\"operationContext\": \"dummyOperationContext\"," +
+                                                "\"resultInfo\": {" +
+                                                "\"code\": 200," +
+                                                "\"subcode\": 200," +
+                                                "\"message\": \"dummyMessage\"" +
+                                                  "}" +
+                                                "}";
+
+        private const string GetCallResultPayload = "{" +
+                                                "\"callConnectionId\": \"af50dd7b-37e7-4cdb-a176-a7c37b71098a\"," +
+                                                "\"targets\": [" +
+                                                   "{" +
+                                                        "\"rawId\": \"dummyRawId\"," +
+                                                        "\"communicationUser\": {" +
+                                                            "\"id\": \"a795d01f-f9ad-45e6-99c4-14bf8449ad4b\"" +
+                                                            "}" +
+                                                    "}," +
+                                                   "{" +
+                                                        "\"rawId\": \"dummyRawId1\"," +
+                                                        "\"phoneNumber\": {" +
+                                                            "\"value\": \"+1555123456\"" +
+                                                            "}" +
+                                                    "}" +
+                                                "]," +
+                                                "\"source\": {" +
+                                                    "\"rawId\": \"dummyRawId\"," +
+                                                    "\"communicationUser\": {" +
+                                                            "\"id\": \"0000000d-5a5f-2db9-ccd7-44482200049a\"" +
+                                                        "}" +
+                                                "}," +
+                                                "\"callConnectionState\": \"connected\"," +
+                                                "\"subject\": \"dummySubject\"," +
+                                                "\"callbackUri\": \"https://callBackUri.local\"," +
+                                                "\"requestedMediaTypes\": [" +
+                                                    "\"audio\"," +
+                                                    "\"video\"" +
+                                                "]," +
+                                                "\"requestedCallEvents\": [" +
+                                                    "\"participantsUpdated\"" +
+                                                "]," +
+                                                "\"callLocator\": {" +
+                                                "\"serverCallId\": \"c07910ba-733a-4848-9e3b-0ec204c8ea34\"," +
+                                                        "\"kind\": \"serverCallLocator\"" +
+                                                    "}" +
+                                                "}";
+
+        private const string GetParticipantsResultPayload = "[" +
+                                                "{" +
+                                                    "\"identifier\": {" +
+                                                        "\"rawId\": \"dummyRawId\"," +
+                                                        "\"communicationUser\": {" +
+                                                                "\"id\": \"0000000d-5a5f-2db9-ccd7-44482200049a\"" +
+                                                            "}" +
+                                                    "}," +
+                                                    "\"participantId\": \"dummyParticipantId\"," +
+                                                    "\"isMuted\": false" +
+                                                "}," +
+                                                "{" +
+                                                    "\"identifier\": {" +
+                                                        "\"rawId\": \"dummyRawId\"," +
+                                                        "\"phoneNumber\": {" +
+                                                                "\"value\": \"+1555123456\"" +
+                                                            "}" +
+                                                    "}," +
+                                                    "\"participantId\": \"dummyParticipantId1\"," +
+                                                    "\"isMuted\": false" +
+                                                "}" +
+                                                "]";
+
+        private const string GetParticipantResultPayload = GetParticipantsResultPayload;
+
+        private const string CreateAudioRoutingGroupResultPayload = "{" +
+                                                                "\"audioRoutingGroupId\": \"dummyAudioRoutingGroupId\"" +
                                                             "}";
 
         [TestCaseSource(nameof(TestData_CallConnectionId))]
@@ -80,8 +153,8 @@ namespace Azure.Communication.CallingServer.Tests
         {
             var callConnection = CreateMockCallConnection(200, CancelAllMediaOperaionsResponsePayload, callConnectionId: callConnectionId);
 
-            var result = await callConnection.CancelAllMediaOperationsAsync().ConfigureAwait(false);
-            VerifyCancelAllMediaOperationsResult(result);
+            var response = await callConnection.CancelAllMediaOperationsAsync().ConfigureAwait(false);
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
         }
 
         [TestCaseSource(nameof(TestData_CallConnectionId))]
@@ -89,8 +162,8 @@ namespace Azure.Communication.CallingServer.Tests
         {
             var callConnection = CreateMockCallConnection(200, CancelAllMediaOperaionsResponsePayload, callConnectionId: callConnectionId);
 
-            var result = callConnection.CancelAllMediaOperations();
-            VerifyCancelAllMediaOperationsResult(result);
+            var response = callConnection.CancelAllMediaOperations();
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
         }
 
         [TestCaseSource(nameof(TestData_CallConnectionId))]
@@ -526,19 +599,19 @@ namespace Azure.Communication.CallingServer.Tests
         [TestCaseSource(nameof(TestData_TransferCall))]
         public async Task TransferCallAsync_Passes(CommunicationIdentifier participant, string targetCallConnectionId, string userToUserInformation)
         {
-            var callConnection = CreateMockCallConnection(202, AddParticipantResultPayload);
+            var callConnection = CreateMockCallConnection(202, TransferCallResultPayload);
 
-            var response = await callConnection.TransferCallAsync(participant, targetCallConnectionId, userToUserInformation).ConfigureAwait(false);
-            Assert.AreEqual((int)HttpStatusCode.Accepted, response.Status);
+            var result = await callConnection.TransferCallAsync(participant, targetCallConnectionId, userToUserInformation).ConfigureAwait(false);
+            VerifyTransferCallResult(result);
         }
 
         [TestCaseSource(nameof(TestData_TransferCall))]
         public void TransferCall_Passes(CommunicationIdentifier participant, string targetCallConnectionId, string userToUserInformation)
         {
-            var callConnection = CreateMockCallConnection(202, AddParticipantResultPayload);
+            var callConnection = CreateMockCallConnection(202, TransferCallResultPayload);
 
-            var response = callConnection.TransferCall(participant, targetCallConnectionId, userToUserInformation);
-            Assert.AreEqual((int)HttpStatusCode.Accepted, response.Status);
+            var result = callConnection.TransferCall(participant, targetCallConnectionId, userToUserInformation);
+            VerifyTransferCallResult(result);
         }
 
         [TestCaseSource(nameof(TestData_TransferCall))]
@@ -557,6 +630,124 @@ namespace Azure.Communication.CallingServer.Tests
             var callConnection = CreateMockCallConnection(404);
 
             RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.TransferCall(participant, targetCallConnectionId, userToUserInformation));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public async Task GetCallAsync_Passes(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(200, GetCallResultPayload, callConnectionId: callConnectionId);
+
+            var result = await callConnection.GetCallAsync().ConfigureAwait(false);
+            VerifyGetCallResult(result);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void GetCall_Passes(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(200, GetCallResultPayload, callConnectionId: callConnectionId);
+
+            var result = callConnection.GetCall();
+            VerifyGetCallResult(result);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void GetCallAsync_Failed(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.GetCallAsync().ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void GetCall_Failed(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.GetCall());
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public async Task GetParticipantsAsync_Passes(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(200, GetParticipantsResultPayload, callConnectionId: callConnectionId);
+
+            var result = await callConnection.GetParticipantsAsync().ConfigureAwait(false);
+            VerifyGetParticipantsResult(result.Value.ToList());
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void GetParticipants_Passes(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(200, GetParticipantsResultPayload, callConnectionId: callConnectionId);
+
+            var result = callConnection.GetParticipants();
+            VerifyGetParticipantsResult(result.Value.ToList());
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void GetParticipantsAsync_Failed(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.GetParticipantsAsync().ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void GetParticipants_Failed(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.GetParticipants());
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public async Task GetParticipantAsync_Passes(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(200, GetParticipantResultPayload, callConnectionId: callConnectionId);
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            var result = await callConnection.GetParticipantAsync(participant).ConfigureAwait(false);
+            VerifyGetParticipantsResult(result.Value.ToList());
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void GetParticipant_Passes(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(200, GetParticipantResultPayload, callConnectionId: callConnectionId);
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            var result = callConnection.GetParticipant(participant);
+            VerifyGetParticipantsResult(result.Value.ToList());
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void GetParticipantAsync_Failed(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.GetParticipantAsync(participant).ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void GetParticipant_Failed(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.GetParticipant(participant));
             Assert.NotNull(ex);
             Assert.AreEqual(ex?.Status, 404);
         }
@@ -607,22 +798,307 @@ namespace Azure.Communication.CallingServer.Tests
             Assert.AreEqual(ex?.Status, 404);
         }
 
-        private void VerifyCancelAllMediaOperationsResult(CancelAllMediaOperationsResult result)
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public async Task KeepAliveAsync_Passes(string callConnectionId)
         {
-            Assert.AreEqual("dummyId", result.OperationId);
-            Assert.AreEqual(OperationStatus.Completed, result.Status);
-            Assert.AreEqual("dummyOperationContext", result.OperationContext);
-            Assert.AreEqual(200, result.ResultInfo.Code);
-            Assert.AreEqual("dummyMessage", result.ResultInfo.Message);
+            var callConnection = CreateMockCallConnection(200, GetParticipantsResultPayload, callConnectionId: callConnectionId);
+
+            var response = await callConnection.KeepAliveAsync().ConfigureAwait(false);
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void KeepAlive_Passes(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(200, GetParticipantsResultPayload, callConnectionId: callConnectionId);
+
+            var response = callConnection.KeepAlive();
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void KeepAliveAsync_Failed(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.KeepAliveAsync().ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_CallConnectionId))]
+        public void KeepAlive_Failed(string callConnectionId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.KeepAlive());
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public async Task HoldParticipantMeetingAudioAsync_Passes(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(200, callConnectionId: callConnectionId);
+
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            var response = await callConnection.HoldParticipantMeetingAudioAsync(participant).ConfigureAwait(false);
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void HoldParticipantMeetingAudio_Passes(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(200, callConnectionId: callConnectionId);
+
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            var response = callConnection.HoldParticipantMeetingAudio(participant);
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void HoldParticipantMeetingAudioAsync_Failed(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.HoldParticipantMeetingAudioAsync(participant).ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void HoldParticipantMeetingAudio_Failed(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.HoldParticipantMeetingAudio(participant));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public async Task ResumeParticipantMeetingAudioAsync_Passes(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(200, callConnectionId: callConnectionId);
+
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            var response = await callConnection.ResumeParticipantMeetingAudioAsync(participant).ConfigureAwait(false);
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void ResumeParticipantMeetingAudio_Passes(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(200, callConnectionId: callConnectionId);
+
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            var response = callConnection.ResumeParticipantMeetingAudio(participant);
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void ResumeParticipantMeetingAudioAsync_Failed(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.ResumeParticipantMeetingAudioAsync(participant).ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_ParticipantId))]
+        public void ResumeParticipantMeetingAudio_Failed(string callConnectionId, string participantUserId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            var participant = new CommunicationUserIdentifier(participantUserId);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.ResumeParticipantMeetingAudio(participant));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_CreateAudioRoutingGroup))]
+        public async Task CreateAudioRoutingGroupAsync_Passes(string callConnectionId, AudioRoutingMode audioRoutingMode, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(201, CreateAudioRoutingGroupResultPayload, callConnectionId: callConnectionId);
+
+            var result = await callConnection.CreateAudioRoutingGroupAsync(audioRoutingMode, targets).ConfigureAwait(false);
+            Assert.IsNotEmpty(result.Value.AudioRoutingGroupId);
+        }
+
+        [TestCaseSource(nameof(TestData_CreateAudioRoutingGroup))]
+        public void CreateAudioRoutingGroup_Passes(string callConnectionId, AudioRoutingMode audioRoutingMode, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(201, CreateAudioRoutingGroupResultPayload, callConnectionId: callConnectionId);
+
+            var result = callConnection.CreateAudioRoutingGroup(audioRoutingMode, targets);
+            Assert.IsNotEmpty(result.Value.AudioRoutingGroupId);
+        }
+
+        [TestCaseSource(nameof(TestData_CreateAudioRoutingGroup))]
+        public void CreateAudioRoutingGroupAsync_Failed(string callConnectionId, AudioRoutingMode audioRoutingMode, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.CreateAudioRoutingGroupAsync(audioRoutingMode, targets).ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_CreateAudioRoutingGroup))]
+        public void CreateAudioRoutingGroup_Failed(string callConnectionId, AudioRoutingMode audioRoutingMode, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.CreateAudioRoutingGroup(audioRoutingMode, targets));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_CreateAudioRoutingGroup))]
+        public void CreateAudioRoutingGroupAsync_FailedWithMode(string callConnectionId, AudioRoutingMode audioRoutingMode, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(404);
+            audioRoutingMode = AudioRoutingMode.OneToOne;
+            ArgumentOutOfRangeException? ex = Assert.ThrowsAsync<ArgumentOutOfRangeException>(async () => await callConnection.CreateAudioRoutingGroupAsync(audioRoutingMode, targets).ConfigureAwait(false));
+            Assert.NotNull(ex);
+        }
+
+        [TestCaseSource(nameof(TestData_CreateAudioRoutingGroup))]
+        public void CreateAudioRoutingGroup_FailedWithMode(string callConnectionId, AudioRoutingMode audioRoutingMode, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(404);
+            audioRoutingMode = AudioRoutingMode.OneToOne;
+            ArgumentOutOfRangeException? ex = Assert.Throws<ArgumentOutOfRangeException>(() => callConnection.CreateAudioRoutingGroup(audioRoutingMode, targets));
+            Assert.NotNull(ex);
+        }
+
+        [TestCaseSource(nameof(TestData_UpdateAudioRoutingGroup))]
+        public async Task UpdateAudioRoutingGroupAsync_Passes(string callConnectionId, string audioRoutingGroupId, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(200, callConnectionId: callConnectionId);
+
+            var response = await callConnection.UpdateAudioRoutingGroupAsync(audioRoutingGroupId, targets).ConfigureAwait(false);
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_UpdateAudioRoutingGroup))]
+        public void UpdateAudioRoutingGroup_Passes(string callConnectionId, string audioRoutingGroupId, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(200, callConnectionId: callConnectionId);
+
+            var response = callConnection.UpdateAudioRoutingGroup(audioRoutingGroupId, targets);
+            Assert.AreEqual((int)HttpStatusCode.OK, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_UpdateAudioRoutingGroup))]
+        public void UpdateAudioRoutingGroupAsync_Failed(string callConnectionId, string audioRoutingGroupId, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.UpdateAudioRoutingGroupAsync(audioRoutingGroupId, targets).ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_UpdateAudioRoutingGroup))]
+        public void UpdateAudioRoutingGroup_Failed(string callConnectionId, string audioRoutingGroupId, IEnumerable<CommunicationIdentifier> targets)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.UpdateAudioRoutingGroup(audioRoutingGroupId, targets));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_DeleteAudioRoutingGroup))]
+        public async Task DeleteAudioRoutingGroupAsync_Passes(string callConnectionId, string audioRoutingGroupId)
+        {
+            var callConnection = CreateMockCallConnection(202, callConnectionId: callConnectionId);
+
+            var response = await callConnection.DeleteAudioRoutingGroupAsync(audioRoutingGroupId).ConfigureAwait(false);
+            Assert.AreEqual((int)HttpStatusCode.Accepted, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_DeleteAudioRoutingGroup))]
+        public void DeleteAudioRoutingGroup_Passes(string callConnectionId, string audioRoutingGroupId)
+        {
+            var callConnection = CreateMockCallConnection(202, callConnectionId: callConnectionId);
+
+            var response = callConnection.DeleteAudioRoutingGroup(audioRoutingGroupId);
+            Assert.AreEqual((int)HttpStatusCode.Accepted, response.Status);
+        }
+
+        [TestCaseSource(nameof(TestData_DeleteAudioRoutingGroup))]
+        public void DeleteAudioRoutingGroupAsync_Failed(string callConnectionId, string audioRoutingGroupId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.ThrowsAsync<RequestFailedException>(async () => await callConnection.DeleteAudioRoutingGroupAsync(audioRoutingGroupId).ConfigureAwait(false));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
+        }
+
+        [TestCaseSource(nameof(TestData_DeleteAudioRoutingGroup))]
+        public void DeleteAudioRoutingGroup_Failed(string callConnectionId, string audioRoutingGroupId)
+        {
+            var callConnection = CreateMockCallConnection(404);
+
+            RequestFailedException? ex = Assert.Throws<RequestFailedException>(() => callConnection.DeleteAudioRoutingGroup(audioRoutingGroupId));
+            Assert.NotNull(ex);
+            Assert.AreEqual(ex?.Status, 404);
         }
 
         private void VerifyPlayAudioResult(PlayAudioResult result)
         {
             Assert.AreEqual("dummyId", result.OperationId);
-            Assert.AreEqual(OperationStatus.Running, result.Status);
+            Assert.AreEqual(CallingOperationStatus.Running, result.Status);
             Assert.AreEqual("dummyOperationContext", result.OperationContext);
             Assert.AreEqual(200, result.ResultInfo.Code);
             Assert.AreEqual("dummyMessage", result.ResultInfo.Message);
+        }
+        private void VerifyTransferCallResult(TransferCallResult result)
+        {
+            Assert.AreEqual("dummyId", result.OperationId);
+            Assert.AreEqual(CallingOperationStatus.Running, result.Status);
+            Assert.AreEqual("dummyOperationContext", result.OperationContext);
+            Assert.AreEqual(200, result.ResultInfo.Code);
+            Assert.AreEqual("dummyMessage", result.ResultInfo.Message);
+        }
+
+        private void VerifyGetCallResult(CallConnectionProperties result)
+        {
+            Assert.NotNull(result);
+            Assert.True(result.Source is CommunicationUserIdentifier);
+            Assert.NotNull(result.Targets);
+            Assert.AreEqual(2, result.Targets.ToList().Count);
+            Assert.True(result.Targets.ElementAt(0) is CommunicationUserIdentifier);
+            Assert.True(result.Targets.ElementAt(1) is PhoneNumberIdentifier);
+            Assert.True(result.CallLocator is ServerCallLocator);
+        }
+
+        private void VerifyGetParticipantResult(List<CallParticipant> result)
+        {
+            VerifyGetParticipantsResult(result);
+        }
+
+        private void VerifyGetParticipantsResult(List<CallParticipant> result)
+        {
+            Assert.NotNull(result);
+            Assert.AreEqual(2, result.Count);
+            Assert.True(result[0].Identifier is CommunicationUserIdentifier);
+            Assert.True(result[1].Identifier is PhoneNumberIdentifier);
         }
 
         private CallConnection CreateMockCallConnection(int responseCode, string? responseContent = null, string callConnectionId = "9ec7da16-30be-4e74-a941-285cfc4bffc5")
@@ -702,6 +1178,7 @@ namespace Azure.Communication.CallingServer.Tests
                 new object?[]
                 {
                     new CommunicationUserIdentifier("8:acs:acsuserid"),
+                    "targetCallConnectionId",
                     "dummyUserToUserInformation"
                 },
             };
@@ -715,6 +1192,52 @@ namespace Azure.Communication.CallingServer.Tests
                 {
                     "d09038e7-38f7-4aa1-9c5c-4bb07a65aa17",
                     "66c76529-3e58-45bf-9592-84eadd52bc81"
+                },
+            };
+        }
+
+        private static IEnumerable<object?[]> TestData_CreateAudioRoutingGroup()
+        {
+            return new[]
+            {
+                new object?[]
+                {
+                    "d09038e7-38f7-4aa1-9c5c-4bb07a65aa17",
+                    AudioRoutingMode.Multicast,
+                    new CommunicationIdentifier[]
+                    {
+                        new CommunicationUserIdentifier("8:acs:resource_target"),
+                        new PhoneNumberIdentifier("+14255550123")
+                    },
+                },
+            };
+        }
+
+        private static IEnumerable<object?[]> TestData_UpdateAudioRoutingGroup()
+        {
+            return new[]
+            {
+                new object?[]
+                {
+                    "d09038e7-38f7-4aa1-9c5c-4bb07a65aa17",
+                    "dummyAudioRoutingGroupId",
+                    new CommunicationIdentifier[]
+                    {
+                        new CommunicationUserIdentifier("8:acs:resource_target"),
+                        new PhoneNumberIdentifier("+14255550123")
+                    },
+                },
+            };
+        }
+
+        private static IEnumerable<object?[]> TestData_DeleteAudioRoutingGroup()
+        {
+            return new[]
+            {
+                new object?[]
+                {
+                    "d09038e7-38f7-4aa1-9c5c-4bb07a65aa17",
+                    "dummyAudioRoutingGroupId",
                 },
             };
         }
